python_services/requirements.txt
fastapi[standard]
pydantic-settings
uvicorn
whoosh
chromadb
junction-python>=0.3.3
yaspin
openai
